Add GPU-aware dispatch, native-unit STL scaling, calibration phantom, DICOM output, performance optimizations, and test suite
- GPU-aware parallel bag generation: workers acquire/release GPUs via shared queue, preventing VRAM contention
- Native-unit STL voxelization: mesh_units parameter converts STL coordinates to mm, mm_per_voxel matches scanner grid
- Configurable scanner factory: create_parallel_scanner() with gantry_diameter_mm and pixel_size_mm
- Calibration phantom config: 5x5 grid of known materials with DICOM ROI metadata for algorithm validation
- DICOM CT output: save_dicom_series() with KVP/spacing metadata, RT-Struct with per-object ROI contours
- Forward model optimization: moved energy loop to numpy/ASTRA-native pipeline, freed torch VRAM for ASTRA projections
- Dynamic Ram-Lak filter: generated from det_col_count instead of hardcoded 1024
- Async FITS I/O: background compression with flush-before-read safety
- stl_only flag: disable primitive shape generation when only STL pool objects are needed
- prevent_overlap flag: lateral shift during placement to reduce object overlap
- Pipeline benchmark: per-stage timing with GPU memory tracking
- Unit tests: 60 tests covering transforms, shapes, FITS I/O, DICOM output
- Fixed np.float deprecation in fdlib.py, numpy .data attribute in cdm_decomposer.py
- Fixed from-numpy-import-* shadowing of builtins (max, min)

+++ lib/misc/fdlib.py
@@ -471,7 +471,7 @@ class wrapped_astra_projector(object):
         self.shape = (
             proj_geom['DetectorCount'] * len(proj_geom['ProjectionAngles']),
             vol_geom['GridColCount'] * vol_geom['GridRowCount'])
-        self.dtype = np.float
+        self.dtype = float
 
     def matvec(self, v):
         sid, s = astra.create_sino(np.reshape(v, (
